+++ lib/configuration.rb
@@ -78,6 +78,7 @@ module AlaveteliConfiguration
       :SMTP_MAILER_AUTHENTICATION => 'plain',
       :SMTP_MAILER_ENABLE_STARTTLS_AUTO => true,
       :SPECIAL_REPLY_VERY_LATE_AFTER_DAYS => 60,
+      :RESTRICT_NEW_RESPONSES_ON_OLD_REQUESTS_AFTER_MONTHS => 6,
       :THEME_BRANCH => false,
       :THEME_URL => "",
       :THEME_URLS => [],